move MessageFilters to laika.config
