+++ core/jvm/src/test/scala/laika/rst/ReStructuredTextToHTMLSpec.scala
@@ -18,11 +18,10 @@ package laika.rst
 
 import cats.data.NonEmptySet
 import laika.api.Transformer
-import laika.api.builder.MessageFilters
 import laika.api.format.TagFormatter
 import laika.ast.Path.Root
 import laika.ast.*
-import laika.config.LaikaKeys
+import laika.config.{ LaikaKeys, MessageFilters }
 import laika.file.FileIO
 import laika.format.{ HTML, ReStructuredText }
 import laika.html.TidyHTML
