+++ core/shared/src/main/scala/laika/config/MessageFilters.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package laika.api.builder
+package laika.config
 
 import laika.ast.MessageFilter
