+++ io/src/test/scala/laika/render/fo/XSLFORendererSpec.scala
@@ -18,7 +18,6 @@ package laika.render.fo
 
 import cats.data.NonEmptySet
 import laika.api.Renderer
-import laika.api.builder.MessageFilters
 import laika.api.bundle.{
   ConfigurablePathTranslator,
   PathAttributes,
@@ -32,7 +31,7 @@ import laika.ast.RelativePath.CurrentTree
 import laika.ast.*
 import laika.ast.sample.{ ParagraphCompanionShortcuts, TestSourceBuilders }
 import laika.ast.styles.{ StyleDeclaration, StyleDeclarationSet, StylePredicate, StyleSelector }
-import laika.config.{ LaikaKeys, TargetFormats }
+import laika.config.{ LaikaKeys, MessageFilters, TargetFormats }
 import laika.format.XSLFO
 import laika.parse.GeneratedSource
 import laika.parse.code.CodeCategory
