+++ sbt/src/main/scala/laika/sbt/Settings.scala
@@ -18,13 +18,13 @@ package laika.sbt
 
 import cats.effect.{ Async, IO, Resource }
 import cats.effect.unsafe.implicits.global
-import laika.api.builder.{ MessageFilters, OperationConfig, ParserBuilder }
+import laika.api.builder.{ OperationConfig, ParserBuilder }
 import laika.api.MarkupParser
 import laika.api.bundle.{ BundleOrigin, ExtensionBundle }
 import laika.api.config.{ Config, ConfigBuilder }
 import laika.api.format.MarkupFormat
 import laika.api.config.Config.ConfigResult
-import laika.config.LaikaKeys
+import laika.config.{ LaikaKeys, MessageFilters }
 import laika.format.{ Markdown, ReStructuredText }
 import laika.io.api.TreeParser
 import laika.io.internal.config.SiteConfig
